First pass at converting datamanager to a module Vue app.
Using SBT vueify here, which is a bit untested but seems to work.
