+++ modules/admin/app/assets/js/components/_comp.vue
@@ -1,0 +1,9 @@
+<script>
+export default {
+
+};
+</script>
+
+<template>
+</template>
+
