+++ modules/admin/app/assets/js/components/log-window.vue
@@ -1,0 +1,15 @@
+<script>
+export default {
+  props: {
+    log: Array,
+  },
+  updated: function () {
+    this.$el.scrollTop = this.$el.scrollHeight;
+  },
+};
+</script>
+
+<template>
+  <pre v-if="log.length > 0"><template v-for="msg in log"><span v-html="msg"></span><br/></template></pre>
+</template>
+
