+++ modules/admin/app/assets/js/components/modal-window.vue
@@ -1,0 +1,27 @@
+<script>
+export default {};
+</script>
+
+<template>
+  <div class="modal show fade" tabindex="-1" role="dialog" style="display: block">
+    <div class="modal-dialog modal-dialog-centered" role="document">
+      <div class="modal-content">
+        <div class="modal-header">
+          <h5 class="modal-title">
+            <slot name="title"></slot>
+          </h5>
+          <button type="button" class="close" aria-label="Close" v-on:click="$emit('close')">
+            <span aria-hidden="true">&times;</span>
+          </button>
+        </div>
+        <div class="modal-body">
+          <slot></slot>
+        </div>
+        <div class="modal-footer">
+          <slot name="footer"></slot>
+        </div>
+      </div>
+    </div>
+  </div>
+</template>
+
