+++ modules/admin/app/assets/js/datamanager-common.js
@@ -533,430 +533,8 @@ Vue.component("modal-alert", {
   `
 });
 
-Vue.component("files-table", {
-  props: {
-    api: DAO,
-    fileStage: String,
-    loadingMore: Boolean,
-    dropping: Boolean,
-    loaded: Boolean,
-    previewing: Object,
-    validating: Object,
-    validationResults: Object,
-    files: Array,
-    selected: Object,
-    truncated: Boolean,
-    deleting: Object,
-    downloading: Object,
-    loadingInfo: Object,
-    filter: String,
-    info: {
-      type: Boolean,
-      default: true
-    }
-  },
-  computed: {
-    allChecked: function () {
-      return Object.keys(this.selected).length === this.files.length;
-    },
-    utilRows: function() {
-      return Number(this.deleted !== null) +
-        Number(this.validating !== null) +
-        Number(this.deleting !== null) +
-        Number(this.downloading !== null) +
-        Number(this.loadingInfo !== null);
-    }
-  },
-  methods: {
-    toggleAll: function (checked) {
-      this.files.forEach(f => this.toggleItem(f, checked));
-    },
-    toggleItem: function (file, checked) {
-      if (checked) {
-        this.$emit('item-selected', file);
-      } else {
-        this.$emit('item-deselected', file);
-      }
-    },
-    isPreviewing: function(file) {
-      return this.previewing !== null && this.previewing.key === file.key;
-    }
-  },
-  watch: {
-    selected: function (newValue) {
-      let selected = Object.keys(newValue).length;
-      let checkAll = this.$el.querySelector("#" + this.fileStage + "-checkall");
-      if (checkAll) {
-        checkAll.indeterminate = selected > 0 && selected !== this.files.length;
-      }
-    },
-  },
-  template: `
-    <div v-bind:class="{'loading': !loaded, 'dropping': dropping}"
-         v-on:keyup.down="$emit('select-next')"
-         v-on:keyup.up="$emit('select-prev')"
-         v-on:click.self.stop="$emit('deselect-all')"
-         class="file-list-container">
-      <table class="table table-bordered table-striped table-sm" v-if="files.length > 0">
-        <thead>
-        <tr>
-          <th><input type="checkbox" v-bind:id="fileStage + '-checkall'" v-on:input="toggleAll($event.target.checked)"/></th>
-          <th>Name</th>
-          <th>Last Modified</th>
-          <th>Size</th>
-          <th v-bind:colspan="utilRows"></th>
-        </tr>
-        </thead>
-        <tbody>
-        <tr v-for="file in files"
-            v-bind:key="file.key"
-            v-on:click.stop="$emit('show-preview', file)"
-            v-bind:class="{'active': isPreviewing(file)}">
-          <td v-on:click.stop.prevent.self="toggleItem(file, !selected[file.key])">
-            <input type="checkbox" v-bind:checked="Boolean(selected[file.key])" 
-                   v-on:input.stop.prevent.self="toggleItem(file, !selected[file.key])"
-                    v-on:click="$event.stopPropagation()">
-          </td>
-          <td>{{file.key|decodeUri}}</td>
-          <td v-bind:title="file.lastModified">{{file.lastModified | prettyDate}}</td>
-          <td>{{file.size | humanFileSize(true)}}</td>
 
-          <td v-if="loadingInfo !== null">
-            <a href="#" v-on:click.prevent.stop="$emit('info', file.key)">
-              <i class="fa fa-fw" v-bind:class="{
-                'fa-circle-o-notch fa-spin': loadingInfo[file.key],
-                'fa-info-circle': !loadingInfo[file.key]
-              }"></i>
-            </a>
-          </td>
-          <td v-if="validating !== null">
-            <a href="#" v-on:click.prevent.stop="$emit('validate-files', _.fromPairs([[file.eTag, file.key]]))">
-                <i v-if="validating[file.eTag]" class="fa fa-fw fa-circle-o-notch fa-spin"></i>
-                <i v-else-if="validationResults && validationResults[file.eTag]" class="fa fa-fw" v-bind:class="{
-                    'fa-check text-success': validationResults[file.eTag].length === 0,
-                    'fa-exclamation-circle text-danger': validationResults[file.eTag].length > 0
-                    }">
-                </i>
-              <i v-else class="fa fa-fw fa-flag-o"></i>
-            </a>
-          </td>
-          <td v-if="downloading !== null">
-            <a href="#" title="" v-on:click.prevent.stop="$emit('download-files', [file.key])">
-              <i class="fa fa-fw" v-bind:class="{
-                'fa-circle-o-notch fa-spin': downloading[file.key],
-                'fa-download': !downloading[file.key]
-              }"></i>
-            </a>
-          </td>
-          <td v-if="deleting !== null">
-            <a href="#" v-on:click.prevent.stop="$emit('delete-files', [file.key])">
-              <i class="fa fa-fw" v-bind:class="{
-                'fa-circle-o-notch fa-spin': deleting[file.key], 
-                'fa-trash-o': !deleting[file.key] 
-              }"></i>
-            </a>
-          </td>
-        </tr>
-        </tbody>
-      </table>
-      <button class="btn btn-sm btn-default" v-if="truncated" v-on:click.prevent.stop="$emit('load-more')">
-        Load more
-        <i v-if="loadingMore" class="fa fa-fw fa-cog fa-spin"/>
-        <i v-else class="fa fa-fw fa-caret-down"/>
-      </button>
-      <div class="panel-placeholder" v-else-if="loaded && filter && files.length === 0">
-        No files found starting with &quot;<code>{{filter}}</code>&quot;...
-      </div>
-      <div class="panel-placeholder" v-else-if="loaded && files.length === 0">
-        There are no files here yet.
-      </div>
-      <div class="file-list-loading-indicator" v-show="!loaded">
-        <i class="fa fa-3x fa-spinner fa-spin"></i>
-      </div>
-    </div>
-  `
-});
 
-Vue.component("file-picker-suggestion", {
-  props: {selected: Boolean, item: Object,},
-  template: `
-    <div @click="$emit('selected', item)" class="file-picker-suggestion" v-bind:class="{'selected': selected}">
-        {{ item.key|decodeUri }} 
-    </div>
-  `
-});
-
-Vue.component("file-picker", {
-  props: {
-    datasetId: String,
-    fileStage: String,
-    value: Object,
-    disabled: Boolean,
-    api: Object,
-    config: Object,
-    placeholder: {
-      type: String,
-      default: "Select file..."
-    }
-  },
-  data: function () {
-    return {
-      text: null,
-      selectedIdx: -1,
-      suggestions: [],
-      loading: false,
-      showSuggestions: false,
-    }
-  },
-  methods: {
-    search: function () {
-      this.loading = true;
-      let list = () => {
-        this.api.listFiles(this.datasetId, this.fileStage, this.text ? this.text : "")
-          .then(data => {
-            this.suggestions = data.files;
-            this.showSuggestions = true;
-          })
-          .finally(() => this.loading = false);
-      }
-      _.debounce(list, 300)();
-    },
-    selectPrev: function () {
-      this.selectedIdx = Math.max(-1, this.selectedIdx - 1);
-    },
-    selectNext: function () {
-      this.selectedIdx = Math.min(this.suggestions.length, this.selectedIdx + 1);
-    },
-    setAndChooseItem: function (item) {
-      this.$emit("input", item);
-      this.cancelComplete();
-      this.text = null;
-    },
-    setItemFromSelection: function () {
-      let idx = this.selectedIdx,
-        len = this.suggestions.length;
-      if (idx > -1 && len > 0 && idx < len) {
-        this.setAndChooseItem(this.suggestions[idx]);
-      } else if (idx === -1) {
-        this.$emit('input', null);
-      }
-    },
-    cancelComplete: function () {
-      this.$nextTick(() => {
-        this.suggestions = [];
-        this.selectedIdx = -1;
-        this.showSuggestions = false;
-      })
-    }
-  },
-  template: `
-    <div class="file-picker">
-      <div class="file-picker-input-container">
-        <div v-show="showSuggestions" class="dropdown-backdrop" v-on:click="cancelComplete"></div>
-        <label class="control-label sr-only">File:</label>
-        <input class="file-picker-input form-control form-control-sm" type="text" v-bind:placeholder="placeholder"
-               v-bind:disabled="disabled"
-               v-bind:value="text !== null ? text : (value ? value.key : '')"
-               v-on:focus="search"
-               v-on:input="text = $event.target.value; search()"
-               v-on:keydown.up="selectPrev"
-               v-on:keydown.down="selectNext"
-               v-on:keydown.enter="setItemFromSelection"
-               v-on:keydown.esc="cancelComplete"/>
-        <i v-if="loading" class="loading-indicator fa fa-circle-o-notch fa-fw fa-spin"></i>
-        <div v-if="showSuggestions" class="file-picker-suggestions dropdown-list">
-          <file-picker-suggestion
-            v-for="(suggestion, i) in suggestions"
-            v-bind:class="{selected: i === selectedIdx}"
-            v-bind:key="suggestion.key"
-            v-bind:item="suggestion"
-            v-bind:selected="i === selectedIdx"
-            v-on:selected="setAndChooseItem"/>
-          <div v-if="!loading && suggestions.length === 0" class="file-picker-suggestions-empty">
-            No files found...
-          </div>
-        </div>
-      </div>
-    </div>
-  `
-});
-
-Vue.component("info-modal", {
-  props: {
-    fileInfo: Object
-  },
-  template: `
-    <modal-window v-on:close="$emit('close')">
-      <template v-slot:title>{{fileInfo.meta.key|decodeUri}}</template>
-      <dl>
-        <dt>File size:</dt>
-        <dd>{{fileInfo.meta.size|humanFileSize}}</dd>
-        <dt>Last Modified:</dt>
-        <dd v-bind:title="fileInfo.meta.lastModified">{{fileInfo.meta.lastModified|prettyDate}}</dd>
-        <dt>E-Tag:</dt>
-        <dd v-bind:title="fileInfo.meta.eTag">{{fileInfo.meta.eTag}}</dd>
-      </dl>
-      <template v-if="fileInfo.versions && fileInfo.versions.length > 1">
-        <h4>Version History</h4>
-        <table class="version-history-list table table-striped table-sm table-bordered">
-          <tr>
-            <th>ID</th>
-            <th>Created</th>
-            <th>Size</th>
-          </tr>
-          <tr v-for="version in fileInfo.versions">
-            <td>{{version.versionId}}</td>
-            <td v-bind:title="version.lastModified">{{version.lastModified|prettyDate}}</td>
-            <td>{{version.size|humanFileSize}}</td>
-          </tr>
-        </table>
-      </template>
-      <template v-slot:footer>
-        <button v-on:click="$emit('close')" type="button" class="btn btn-default">
-          Close
-        </button>
-      </template>
-    </modal-window>
-  `
-});
-
-Vue.component("filter-control", {
-  props: {
-    filter: Object
-  },
-  template: `
-    <div class="filter-control">
-      <label class="sr-only">Filter files</label>
-      <input class="filter-input form-control form-control-sm" type="text" v-model.trim="filter.value"
-             placeholder="Filter files..." v-on:keyup="$emit('filter')"/>
-      <i class="filtering-indicator fa fa-circle-o-notch fa-fw fa-spin" v-if="filter.active"/>
-      <i class="filtering-indicator fa fa-close fa-fw" style="cursor: pointer" v-on:click="$emit('clear')" v-else-if="filter.value"/>
-    </div>
-  `
-});
-
-Vue.component("validate-button", {
-  props: {
-    disabled: Boolean,
-    selected: Number,
-    active: Boolean,
-  },
-  template: `
-    <button v-bind:disabled="disabled" v-on:click.prevent="$emit('validate')" class="btn btn-sm btn-default">
-      <i class="fa fa-fw" v-bind:class="{'fa-flag-o': !active, 'fa-circle-o-notch fa-spin': active}"/>
-      <template v-if="selected > 0">
-        Validate Selected ({{selected}})
-      </template>
-      <template v-else>
-        Validate All
-      </template>
-    </button>
- `
-});
-
-Vue.component("delete-button", {
-  props: {
-    disabled: Boolean,
-    selected: Number,
-    active: Boolean,
-  },
-  template: `
-    <button v-bind:disabled="disabled" v-on:click.prevent="$emit('delete')" class="btn btn-sm btn-default">
-      <i class="fa fa-fw" v-bind:class="{'fa-trash-o': !active, 'fa-circle-o-notch fa-spin': active}"/>
-      <template v-if="selected > 0">
-        Delete Selected ({{selected}})
-      </template>
-      <template v-else>
-        Delete All
-      </template>
-    </button>
-  `
-})
-
-Vue.component("log-window", {
-  props: {
-    log: Array,
-  },
-  updated: function () {
-    this.$el.scrollTop = this.$el.scrollHeight;
-  },
-  template: `
-    <pre v-if="log.length > 0"><template v-for="msg in log"><span v-html="msg"></span><br/></template></pre>
-  `
-});
-
-Vue.component("drag-handle", {
-  props: {
-    ns: String,
-    p2: Element,
-    container: Element,
-  },
-  data: function () {
-    return {
-      offset: 0,
-    }
-  },
-
-  methods: {
-    move: function (evt) {
-      // Calculate the height of the topmost panel in percent.
-      let maxY = this.container.offsetTop + this.container.offsetHeight;
-      let topY = this.container.offsetTop;
-      let posY = evt.clientY - this.offset;
-
-      let pxHeight = Math.min(maxY, Math.max(0, posY - topY));
-      let percentHeight = pxHeight / this.container.offsetHeight * 100;
-
-      // Now convert to the height of the lower panel.
-      let perc = 100 - percentHeight;
-      this.p2.style.flexBasis = perc + "%";
-    },
-    startDrag: function (evt) {
-      let us = this.container.style.userSelect;
-      let cursor = this.container.style.cursor;
-      this.offset = evt.clientY - this.$el.offsetTop;
-      this.container.addEventListener("mousemove", this.move);
-      this.container.style.userSelect = "none";
-      this.container.style.cursor = "ns-resize";
-      window.addEventListener("mouseup", () => {
-        console.log("Stop resize");
-        this.offset = 0;
-        this.$emit("resize", this.p2.clientHeight);
-        this.container.style.userSelect = us;
-        this.container.style.cursor = cursor;
-        this.container.removeEventListener("mousemove", this.move);
-      }, {once: true});
-    },
-  },
-  template: `
-    <div v-bind:id="ns + '-drag-handle'" class="drag-handle" v-on:mousedown="startDrag"></div>
-  `
-});
-
-Vue.component("modal-window", {
-  template: `
-    <div class="modal show fade" tabindex="-1" role="dialog" style="display: block">
-      <div class="modal-dialog modal-dialog-centered" role="document">
-        <div class="modal-content">
-          <div class="modal-header">
-            <h5 class="modal-title">
-                <slot name="title"></slot>
-            </h5>
-            <button type="button" class="close" aria-label="Close" v-on:click="$emit('close')">
-              <span aria-hidden="true">&times;</span>
-            </button>
-          </div>
-          <div class="modal-body">
-            <slot></slot>
-          </div>
-          <div class="modal-footer">
-            <slot name="footer"></slot>
-          </div>
-        </div>
-      </div>
-    </div>
-  `
-});
 
 
 let initialStageState = function() {
