+++ package.json
@@ -32,8 +32,18 @@
   "author": "EHRI",
   "license": "EUPL-1.2",
   "devDependencies": {
+    "@vue/test-utils": "^1.1.3",
     "grunt": "^1.3.0",
-    "grunt-contrib-copy": "^1.0.0"
+    "grunt-contrib-copy": "^1.0.0",
+    "jest": "^26.6.3",
+    "jsdom": "^16.4.0",
+    "jsdom-global": "^3.0.2",
+    "ts-jest": "^26.5.3",
+    "typescript": "^4.2.2",
+    "vue-loader": "^15.9.6",
+    "vue-template-compiler": "^2.6.12",
+    "webpack": "^5.24.3",
+    "webpack-cli": "^4.5.0"
   },
   "repository": {
     "type": "git",
