+++ project/plugins.sbt
@@ -21,3 +21,7 @@ addSbtPlugin("net.ground5hark.sbt" % "sbt-concat" % "0.2.0")
 
 // Excluded because too slow for workable development
 //addSbtPlugin("net.ground5hark.sbt" % "sbt-css-compress" % "0.1.4")
+
+resolvers += Resolver.bintrayRepo("givers", "maven")
+
+addSbtPlugin("givers.vuefy" % "sbt-vuefy" % "6.0.0")
